requirements.txt
django-heroku==0.3.1
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
yapf==0.30.0
django-cors-headers==3.7.0
djangorestframework
